daemon: reexecute_blocks_on_sync

+++ xelis_daemon/src/p2p/chain_sync.rs
@@ -211,7 +211,7 @@ impl<S: Storage> P2pServer<S> {
                 let block = Block::new(Immutable::Arc(header), transactions);
                 // don't broadcast block because it's syncing
                 self.blockchain.add_new_block(block, Some(Immutable::Owned(hash)), BroadcastOption::Miners, false).await?;
-            } else {
+            } else if self.reexecute_blocks_on_sync {
                 // We need to re execute it to make sure it's in DAG
                 let mut storage = self.blockchain.get_storage().write().await;
                 if !storage.is_block_topological_ordered(&hash).await? {
@@ -522,34 +522,36 @@ impl<S: Storage> P2pServer<S> {
                                         }
                                     },
                                     ResponseHelper::NotRequested(hash) => {
-                                        let is_ordered = {
-                                            debug!("locking storage for block ordering check");
-                                            let storage = self.blockchain.get_storage().read().await;
-                                            debug!("storage read acquired for block ordering check");
-                                            storage.is_block_topological_ordered(&hash).await?
-                                        };
-
-                                        // Block is not ordered, we may have an issue, we should
-                                        // force its re execution, for that we delete it from our storage
-                                        if !is_ordered {
-                                            warn!("Forcing block {} re-execution", hash);
-                                            let mut storage = self.blockchain.get_storage().write().await;
-                                            debug!("storage write acquired for block forced re-execution");
-
-                                            let block = storage.delete_block_with_hash(&hash).await?;
-                                            let mut tips = storage.get_tips().await?;
-                                            if tips.remove(&hash) {
-                                                debug!("Block {} was a tip, removing it from tips", hash);
-                                                storage.store_tips(&tips).await?;
+                                        if self.reexecute_blocks_on_sync {
+                                            let is_ordered = {
+                                                debug!("locking storage for block ordering check");
+                                                let storage = self.blockchain.get_storage().read().await;
+                                                debug!("storage read acquired for block ordering check");
+                                                storage.is_block_topological_ordered(&hash).await?
+                                            };
+    
+                                            // Block is not ordered, we may have an issue, we should
+                                            // force its re execution, for that we delete it from our storage
+                                            if !is_ordered {
+                                                warn!("Forcing block {} re-execution", hash);
+                                                let mut storage = self.blockchain.get_storage().write().await;
+                                                debug!("storage write acquired for block forced re-execution");
+    
+                                                let block = storage.delete_block_with_hash(&hash).await?;
+                                                let mut tips = storage.get_tips().await?;
+                                                if tips.remove(&hash) {
+                                                    debug!("Block {} was a tip, removing it from tips", hash);
+                                                    storage.store_tips(&tips).await?;
+                                                }
+    
+                                                // Replicate same behavior as above branch
+                                                if let Err(e) = self.blockchain.add_new_block_for_storage(&mut storage, block, Some(Immutable::Owned(hash)), BroadcastOption::Miners, false).await {
+                                                    drop(futures);
+    
+                                                    self.object_tracker.mark_group_as_fail(group_id).await;
+                                                    return Err(e)
+                                                }                                            
                                             }
-
-                                            // Replicate same behavior as above branch
-                                            if let Err(e) = self.blockchain.add_new_block_for_storage(&mut storage, block, Some(Immutable::Owned(hash)), BroadcastOption::Miners, false).await {
-                                                drop(futures);
-
-                                                self.object_tracker.mark_group_as_fail(group_id).await;
-                                                return Err(e)
-                                            }                                            
                                         }
                                     }
                                 },
@@ -581,7 +583,7 @@ impl<S: Storage> P2pServer<S> {
                             return Err(P2pError::ExpectedBlock(response).into())
                         }
                         total_requested += 1;
-                    } else {
+                    } else if self.reexecute_blocks_on_sync {
                         trace!("Block {} is already in chain, verify if its in DAG", hash);
 
                         let block = {
